fix: int32 to int64

+++ src/Backend/MyMusicLibrary.Infrastructure/DataAccess/Migrations/Versions/Version11.cs
@@ -10,7 +10,7 @@ namespace MyMusicLibrary.Migrations
         public override void Up()
         {
             Create.Table("UserFavoritesMusic")
-                .WithColumn("Id").AsInt32().PrimaryKey().Identity()
+                .WithColumn("Id").AsInt64().PrimaryKey().Identity()
                 .WithColumn("UserId").AsInt64().NotNullable()
                 .WithColumn("MusicId").AsInt64().NotNullable()
                 .WithColumn("Active").AsBoolean().NotNullable().WithDefaultValue(true) // adiciona Active
